Move code to run/env_vars.rs
