+++ evobench-evaluator/src/run/output_directory_structure.rs
@@ -15,7 +15,7 @@ use crate::{
     ctx,
     git::GitHash,
     key::RunParameters,
-    run::run_job::AllowableCustomEnvVar,
+    run::env_vars::AllowableCustomEnvVar,
     serde::{
         allowed_env_var::AllowedEnvVar, date_and_time::DateTimeWithOffset,
         proper_dirname::ProperDirname, proper_filename::ProperFilename,
